도서 전체 조회 기능

+++ controller/cartController.js
@@ -1,8 +1,7 @@
+const ensureAuthorization = require('../auth')
 const jwt = require('jsonwebtoken')
 const conn = require('../mariadb');
 const { StatusCodes } = require('http-status-codes');
-const dotenv = require('dotenv');
-dotenv.config();
 
 const addToCart = (req, res) => {
     const { book_id, quantity } = req.body;
@@ -49,8 +48,14 @@ const getCartItems = (req, res) => {
         let sql = `SELECT cartItems.id, book_id, title, summary, quantity, price 
                 FROM cartItems LEFT JOIN books 
                 ON cartItems.book_id = books.id
-                WHERE user_id = ? AND cartItems.id IN (?)`;
-        let values = [authorization.id, selected];
+                WHERE user_id = ? `;
+        let values = [authorization.id];
+
+        if (selected) { // 주문서 작성 시 선택한 장바구니 목록 조회
+            sql += ` AND cartItems.id IN (?)`;
+            values.push(selected);
+        }
+
         conn.query(sql, values,
             (err, results) => {
                 if (err) {
@@ -64,34 +69,30 @@ const getCartItems = (req, res) => {
 };
 
 const removeCartItem = (req, res) => {
-    const id = req.params.id;
+    let authorization = ensureAuthorization(req);
 
-    let sql = "DELETE FROM cartItems WHERE id = ?";
-    conn.query(sql, id,
-        (err, results) => {
-            if (err) {
-                console.log(err);
-                return res.status(StatusCodes.BAD_REQUEST).end();
-            }
+    if (authorization instanceof jwt.TokenExpiredError) {
+        return res.status(StatusCodes.UNAUTHORIZED).json({
+            message: "로그인 세션이 만료되어 재로그인이 필요합니다"
+        });
+    } else if (authorization instanceof jwt.JsonWebTokenError) {
+        return res.status(StatusCodes.BAD_REQUEST).json({
+            message: "잘못된 토큰 입니다."
+        });
+    } else {
+        const id = req.params.id;
 
-            return res.status(StatusCodes.CREATED).json(results);
-        })
-};
+        let sql = "DELETE FROM cartItems WHERE id = ?";
+        conn.query(sql, id,
+            (err, results) => {
+                if (err) {
+                    console.log(err);
+                    return res.status(StatusCodes.BAD_REQUEST).end();
+                }
 
-function ensureAuthorization(req) {
-    try {
-        let receivedJwt = req.headers["authorization"];
-        console.log(receivedJwt)
-
-        let decodedJWT = jwt.verify(receivedJwt, process.env.PRIVATE_kEY);
-        console.log(decodedJWT);
-
-        return decodedJWT;
-    } catch (err) {
-        console.log(err.name + " : " + err.message);
-
-        return err;
+                return res.status(StatusCodes.CREATED).json(results);
+            })
     }
-}
+};
 
 module.exports = { addToCart, getCartItems, removeCartItem };
